A script to parse .ccj files from Crossword Compiler
This format is used by many newspapers to publish their online
crosswords, either allowing you to download the file to solve
off-line or as input for a Java applet.

This script can also generate something like an Across Lite binary
.PUZ file which can be loaded into xword but almost certainly not
Across Lite itself.  There are notes in the script indicating what
to do if you mind about this.