Don't modify the clue lists on writing out in .puz format
We sometimes need to generate extra clues before writing the
.puz format (when a clue spans several entries) - previously
the clue lists stored by the ParsedCCJ were modified on doing
this, which is obviously undesirable.  This commit changes
that behaviour so that the extra clues are only added to a
copy of the clue lists generated for the purposes of creating
the output.